Fixed issue with player walking on units
Previously the positions of all units and the player were tracked on a 2D array. When the player walked onto the same tile as a unit it would overwrite the unit currently in that spot, removing the unit from the array and preventing it from being tracked for updates. This commit fixes the issue by preventing the player from writing itself to the 2D array.