Added health system
Units now have health and attacks deal damage to a units health. Each unit is given 4 health and die when their health reaches 0. If a unit attacks and wins, it loses no health and reduces the enemies health by 1. If the unit attacks and its a stalemate, both sides take 1 damage.